Refactor the GameEstimator to move it into the API
- Remove references to Params from the GameEstimator
- Make the members of the GameEstimator similar to those of the spark.ml.Estimator
- Refactor and simplify the GameEstimator integration tests
- Make several API level constructs accessible from anywhere (not just within Photon-ML)
