--- photon-client/src/integTest/scala/com/linkedin/photon/ml/event/TestEventListener.scala
@@ -1,48 +1,0 @@
-/*
- * Copyright 2017 LinkedIn Corp. All rights reserved.
- * Licensed under the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License. You may obtain a
- * copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-package com.linkedin.photon.ml.event
-
-/**
- * Test [[EventListener]] for [[EventListenerTest]].
- */
-class TestEventListener extends EventListener {
-  import TestEventListener._
-
-  private var valueStore: Int = DEFAULT_VALUE
-
-  /**
-   *
-   * @return
-   */
-  def getValue: Int = valueStore
-
-  /**
-   *
-   * @param event The event
-   */
-  override def handle(event: Event): Unit = event match {
-    case e: TestEvent => valueStore = e.value
-    case _ =>
-  }
-
-  /**
-   *
-   */
-  override def close(): Unit = valueStore = DEFAULT_VALUE
-}
-
-object TestEventListener {
-  val DEFAULT_VALUE = 0
-}
