+++ photon-client/src/main/scala/com/linkedin/photon/ml/cli/game/training/Driver.scala
@@ -22,30 +22,31 @@ import org.apache.spark.sql.DataFrame
 import org.apache.spark.sql.functions.col
 import org.slf4j.Logger
 
-import com.linkedin.photon.ml.Types.{FeatureShardId, SparkVector}
+import com.linkedin.photon.ml.Types._
 import com.linkedin.photon.ml.cli.game.GameDriver
+import com.linkedin.photon.ml.data.InputColumnsNames
 import com.linkedin.photon.ml.data.avro.{AvroDataReader, ModelProcessingUtils}
-import com.linkedin.photon.ml.estimators.{GameEstimator, GameParams}
+import com.linkedin.photon.ml.estimators.GameEstimator
 import com.linkedin.photon.ml.evaluation.Evaluator.EvaluationResults
 import com.linkedin.photon.ml.io.deprecated.ModelOutputMode
 import com.linkedin.photon.ml.model.GameModel
-import com.linkedin.photon.ml.normalization.{NormalizationContext, NormalizationType}
+import com.linkedin.photon.ml.normalization.{NormalizationType, NormalizationContext}
 import com.linkedin.photon.ml.optimization.game.GameModelOptimizationConfiguration
 import com.linkedin.photon.ml.stat.BasicStatisticalSummary
 import com.linkedin.photon.ml.util.Implicits._
-import com.linkedin.photon.ml.util.Utils._
+import com.linkedin.photon.ml.util.Utils
 import com.linkedin.photon.ml.util._
-import com.linkedin.photon.ml.{Constants, InputColumnsNames, SparkContextConfiguration}
+import com.linkedin.photon.ml.{Constants, SparkContextConfiguration}
 
 /**
  * The Driver class, which drives the training of GAME model.
  *
  * @note there is a separate Driver to drive the scoring of GAME models.
  */
-final class Driver(val sc: SparkContext, val params: GameParams, implicit val logger: Logger)
+final class Driver(val sc: SparkContext, val params: GameTrainingParams, implicit val logger: Logger)
   extends GameDriver(sc, params, logger) {
 
-  // These two types make the code easier to read, and are somewhat specific to the GAME Driver
+  // These types make the code easier to read, and are somewhat specific to the GAME Driver
   type FeatureShardStatistics = Iterable[(FeatureShardId, BasicStatisticalSummary)]
   type FeatureShardStatisticsOpt = Option[FeatureShardStatistics]
   type IndexMapLoaders = Map[FeatureShardId, IndexMapLoader]
@@ -75,10 +76,34 @@ final class Driver(val sc: SparkContext, val params: GameParams, implicit val lo
 
     val normalizationContexts = Timed("Prepare normalization contexts") {
       prepareNormalizationContexts(trainingData, featureIndexMapLoaders, featureShardStats)
+        .map(_.mapValues(context => sc.broadcast(context)))
     }
 
-    val models = Timed("Fit model") {
-      new GameEstimator(sc, params, logger).fit(trainingData, validationData, normalizationContexts)
+    val models = Timed("Fit models") {
+      val estimator = new GameEstimator(sc, logger)
+        .setDatumInputColumnNames(params.inputColumnsNames)
+        .setFeatureShardColumnNames(params.featureShardIdToFeatureSectionKeysMap.keys.toSet)
+        .setTaskType(params.taskType)
+        .setUpdatingSequence(params.updatingSequence)
+        .setFixedEffectDataConfigurations(params.fixedEffectDataConfigurations)
+        .setRandomEffectDataConfigurations(params.randomEffectDataConfigurations)
+        .setNumOuterIterations(params.numIterations)
+        .setComputeVariance(params.computeVariance)
+        .setEvaluatorTypes(params.evaluatorTypes)
+        .setNormalizationContexts(normalizationContexts)
+
+      val modelConfigs = for (
+        fixedEffectOptimizationConfiguration <- params.fixedEffectOptimizationConfigurations;
+        randomEffectOptimizationConfiguration <- params.randomEffectOptimizationConfigurations;
+        factoredRandomEffectOptimizationConfiguration <- params.factoredRandomEffectOptimizationConfigurations) yield {
+
+        GameModelOptimizationConfiguration(
+          fixedEffectOptimizationConfiguration,
+          randomEffectOptimizationConfiguration,
+          factoredRandomEffectOptimizationConfiguration)
+      }
+
+      estimator.fit(trainingData, validationData, modelConfigs)
     }
 
     val bestModel = Timed("Select best model") {
@@ -106,7 +131,6 @@ final class Driver(val sc: SparkContext, val params: GameParams, implicit val lo
    *
    * @note this returns an Option rather than a naked DataFrame to harmonize the signature with readValidationData.
    *       readTrainingData will most probably always return a "full" Option.
-   *
    * @param featureIndexMapLoaders The feature index map loaders
    * @return An Option containing the loaded data frame
    */
@@ -132,11 +156,13 @@ final class Driver(val sc: SparkContext, val params: GameParams, implicit val lo
     val numPartitions = math.max(numFixedEffectPartitions, numRandomEffectPartitions)
     require(numPartitions > 0, "Invalid configuration: neither fixed effect nor random effect partitions specified.")
 
-    Some(new AvroDataReader(sc).readMerged(
-      trainingRecordsPath,
-      featureIndexMapLoaders,
-      params.featureShardIdToFeatureSectionKeysMap,
-      numPartitions))
+    Some(
+      new AvroDataReader(sc)
+        .readMerged(
+          trainingRecordsPath,
+          featureIndexMapLoaders,
+          params.featureShardIdToFeatureSectionKeysMap,
+          numPartitions))
   }
 
   /**
@@ -167,21 +193,17 @@ final class Driver(val sc: SparkContext, val params: GameParams, implicit val lo
     }
 
   /**
-   * Check data, i.e. verify that it is admissible for training and/or validation.
+   * Check data, i.e. verify that it is admissible for training and/or validation. Currently, this checks that each
+   * sample weight has a strictly non-zero weight.
    *
-   * For now, we just check that the data contains at least a sample with non-zero weight). We throw out samples that
-   * have a weight less than or equal to 0. If we don't find at least one sample with a strictly positive weight, we
-   * throw an IllegalArgumentException, otherwise we return the filtered data set.
-   *
-   * @note this function has the side-effect of possibly reducing the size of the data set
-   * @note this is somewhat expensive (it iterates over the whole data), so the user can control it via parameter
-   *       checkData
-   *
+   * @note This is somewhat expensive (it iterates over the whole data), so the user can control it via the parameter
+   *       'checkData'
+   * @note We have explicitly chosen to 'verify and reject' rather than 'identify and fix' - we feel that it will ease
+   *       the burden of maintaining the code
    * @param data The data to check
-   * @return A data set where zero-weight samples have been removed, or an exception if we couldn't find at least one
-   *         sample with a strictly positive weight
+   * @throws IllegalArgumentException if at least one sample with a strictly positive weight could not be found
    */
-  protected[training] def checkData(data: Option[DataFrame]): Option[DataFrame] = {
+  protected[training] def checkData(data: Option[DataFrame]): Unit = {
 
     val weightColumnName = params.inputColumnsNames(InputColumnsNames.WEIGHT)
 
@@ -195,8 +217,8 @@ final class Driver(val sc: SparkContext, val params: GameParams, implicit val lo
   }
 
   /**
-   * Helper to avoid writing the same code for training and validation data: read some data set, and then optionally
-   * validate it, while timing each step separately.
+   * Helper to avoid writing the same code for training and validation data: read some data set, then optionally
+   * validate it, timing each step separately.
    *
    * @param reader A function that will read a data set to an Option[DataFrame]
    * @param dataName The name of the data set to read and check ("training", "validation"...)
@@ -205,31 +227,42 @@ final class Driver(val sc: SparkContext, val params: GameParams, implicit val lo
    */
   protected[training] def readAndCheck(reader: => Option[DataFrame], dataName: String): Option[DataFrame] = {
 
-    val tmp = Timed(s"Read $dataName") { reader }
+    val dataframe = Timed(s"Read $dataName") { reader }
 
-    if (params.checkData)
-      Timed(s"Check $dataName") { checkData(tmp) }
-    else
-      tmp
+    if (params.checkData) {
+      Timed(s"Check $dataName") {
+        checkData(dataframe)
+      }
+    }
+
+    dataframe
   }
 
   /**
-   * Calculate basic statistics (same as spark-ml) on a DataFrame.
+   * Prepare normalization contexts, if normalization is enabled.
    *
-   * @param data The data to compute statistics on
-   * @param featureIndexMapLoaders The index map loaders to use to retrieve the feature shards
-   * @return One BasicStatisticalSummary per feature shard
+   * @param trainingData The training data
+   * @param featureIndexMapLoaders The index map loaders
+   * @return Normalization contexts for each coordinate, or None if normalization is disabled
    */
-  private def calculateStatistics(
-      data: DataFrame,
-      featureIndexMapLoaders: IndexMapLoaders): FeatureShardStatistics =
+  protected def prepareNormalizationContexts(
+    trainingData: DataFrame,
+    featureIndexMapLoaders: IndexMapLoaders,
+    statistics: FeatureShardStatisticsOpt): Option[Map[CoordinateId, NormalizationContext]] =
 
-    featureIndexMapLoaders.keys
-      .map {
-        featureShardId =>
-          // Calling rdd explicitly here to avoid a typed encoder lookup in Spark 2.1
-          (featureShardId, BasicStatisticalSummary(data.select(featureShardId).rdd.map(_.getAs[SparkVector](0))))
-      }
+    Utils.filter(params.normalizationType != NormalizationType.NONE) {
+      val featureShardToNormalizationContextMap = statistics
+        .getOrElse(calculateStatistics(trainingData, featureIndexMapLoaders.keys))
+        .map { case (featureShardId, featureShardStats) =>
+          val intercept = featureIndexMapLoaders(featureShardId).indexMapForDriver().get(Constants.INTERCEPT_KEY)
+          (featureShardId, NormalizationContext(params.normalizationType, featureShardStats, intercept))
+        }
+        .toMap
+      val coordinateIdToFeatureShardMap = params.fixedEffectDataConfigurations.mapValues(_.featureShardId) ++
+        params.randomEffectDataConfigurations.mapValues(_.featureShardId)
+
+      coordinateIdToFeatureShardMap.mapValues(featureShardId => featureShardToNormalizationContextMap(featureShardId))
+    }
 
   /**
    * Compute basic statistics (same as spark-ml) of the training data for each feature shard.
@@ -244,37 +277,30 @@ final class Driver(val sc: SparkContext, val params: GameParams, implicit val lo
       featureIndexMapLoaders: IndexMapLoaders): FeatureShardStatisticsOpt =
 
     params.summarizationOutputDirOpt
-      .map {
-        (summarizationOutputDir: String) => {
-          calculateStatistics(trainingData, featureIndexMapLoaders)
-            .tap { case (featureShardId, featureShardStats) =>
-              val outputDir = summarizationOutputDir + "/" + featureShardId
-              val indexMap = featureIndexMapLoaders(featureShardId).indexMapForDriver()
-              ModelProcessingUtils.writeBasicStatistics(sc, featureShardStats, outputDir, indexMap)
-            }
-        }
+      .map { summarizationOutputDir: String =>
+        calculateStatistics(trainingData, featureIndexMapLoaders.keys)
+          .tap { case (featureShardId, featureShardStats) =>
+            val outputDir = summarizationOutputDir + "/" + featureShardId
+            val indexMap = featureIndexMapLoaders(featureShardId).indexMapForDriver()
+            ModelProcessingUtils.writeBasicStatistics(sc, featureShardStats, outputDir, indexMap)
+          }
       }
 
   /**
-   * Prepare normalization contexts, if the normalization options has been setup in the parameters.
+   * Calculate basic statistics (same as spark-ml) on a DataFrame.
    *
-   * @param trainingData The training data
-   * @param featureIndexMapLoaders The index map loaders
-   * @return Normalization contexts for each featureShardId, or None if normalization is not needed
+   * @param data The data to compute statistics on
+   * @param featureShards The feature shards for which to compute statistics
+   * @return One BasicStatisticalSummary per feature shard
    */
-  protected[training] def prepareNormalizationContexts(
-      trainingData: DataFrame,
-      featureIndexMapLoaders: IndexMapLoaders,
-      statistics: FeatureShardStatisticsOpt): Option[Map[FeatureShardId, NormalizationContext]] =
+  protected def calculateStatistics(
+    data: DataFrame,
+    featureShards: Iterable[FeatureShardId]): FeatureShardStatistics =
 
-    Filter(params.normalizationType != NormalizationType.NONE) {
-      statistics
-        .getOrElse(calculateStatistics(trainingData, featureIndexMapLoaders))
-        .map { case (featureShardId, featureShardStats) =>
-          val intercept = featureIndexMapLoaders(featureShardId).indexMapForDriver().get(Constants.INTERCEPT_KEY)
-          (featureShardId, NormalizationContext(params.normalizationType, featureShardStats, intercept))
-        }
-        .toMap
+    featureShards.map {
+      featureShardId =>
+        // Calling rdd explicitly here to avoid a typed encoder lookup in Spark 2.1
+        (featureShardId, BasicStatisticalSummary(data.select(featureShardId).rdd.map(_.getAs[SparkVector](0))))
     }
 
   /**
@@ -384,7 +410,7 @@ object Driver {
    */
   def main(args: Array[String]): Unit = {
 
-    val tryParams = Try(GameParams.parseFromCommandLine(args)) // An exception can be thrown by parseFromCommandLine
+    val tryParams = Try(GameTrainingParams.parseFromCommandLine(args)) // An exception can be thrown by parseFromCommandLine
 
     tryParams match {
