+++ photon-test-utils/build.gradle
@@ -21,6 +21,8 @@ dependencies {
   compile("org.apache.spark:spark-core$scalaSuffix:$sparkVersion") {
     exclude group: 'org.apache.avro', module: 'avro-mapred'
   }
+  compile("org.apache.spark:spark-sql$scalaSuffix:$sparkVersion")
+
   compile("org.scalanlp:breeze$scalaSuffix:0.11.2")
   compile("org.scalanlp:breeze-macros$scalaSuffix:0.11.2")
