+++ photon-test-utils/src/main/scala/com/linkedin/photon/ml/test/SparkTestUtils.scala
@@ -21,7 +21,8 @@ import breeze.linalg.{SparseVector, Vector}
 import org.apache.commons.math3.distribution.PascalDistribution
 import org.apache.commons.math3.random.{RandomGenerator, Well19937a}
 import org.apache.log4j.{LogManager, Logger}
-import org.apache.spark.{SparkConf, SparkContext}
+import org.apache.spark.SparkContext
+import org.apache.spark.sql.SparkSession
 
 // TODO: Additional documentation required
 /**
@@ -29,6 +30,7 @@ import org.apache.spark.{SparkConf, SparkContext}
  */
 trait SparkTestUtils {
 
+  var spark: SparkSession = _
   var sc: SparkContext = _
 
   /**
@@ -59,14 +61,14 @@ trait SparkTestUtils {
   def sparkTest(name: String)(body: => Unit): Unit = {
     SparkTestUtils.SPARK_LOCAL_CONFIG.synchronized {
 
-      val conf: SparkConf = new SparkConf()
-      conf.setAppName(name).setMaster(SparkTestUtils.SPARK_LOCAL_CONFIG)
-      sc = new SparkContext(conf)
+      spark = SparkSession.builder().master(SparkTestUtils.SPARK_LOCAL_CONFIG).appName(name).getOrCreate()
+      sc = spark.sparkContext
 
       try {
         body
       } finally {
         sc.stop()
+        spark.stop()
         System.clearProperty("spark.driver.port")
         System.clearProperty("spark.hostPort")
       }
